requirements.txt
quart
aiosqlite
PyYaml
revChatGPT
tiktoken
claude-api
bardapi
hugchat
g4f
revTongYi
colorlog
unofficial-claude2-api